changed SDL2 to SDL and implimented more of the interfaces. some methods not complete though.
